Phase 1.6777777777: reports in HTML
Summary builder it’s alive, and it does summaries, and we use them to
generate reports in the views.

PDF generation, in the other hand, is failing, hard. Gave up on using
thin records and using wicked_pdf to generate the files from plain HTML
views, and generating garbage for now.

Except for that it’s almost good. Next phase will be to butcher the
models (already started but with stubs), to add new fields needed for
the reports, but still need to fix the PDFs.